Created initial benchmark.

+++ apps/ising_spin_fix/create_ising_spin_fix_instance.py
@@ -49,7 +49,7 @@ for i in range(0,5):
         sys.stderr.write("prob[{}] = {} = p\n".format(index,probs[index], p))
 
 properties={
-    "endTime":endTime * 2**20,
+    "endTime":int(endTime * 2**20),
     "width":n,
     "height":n,
     "probabilities":probs
